@@ -2,7 +2,7 @@
 version = "0.0.0"                               # managed by cargo-workspaces, see below
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2021"
-rust-version = "1.80.1"
+rust-version = "1.81.0"
 repository = "https://github.com/near/nearcore"
 license = "MIT OR Apache-2.0"
 
